feat: extract the useDragNode hook
